requirements.txt
insightface==0.7.3
onnxruntime==1.17.3
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
scikit-learn==1.4.2
hdbscan==0.8.38.post1
ffmpeg-python==0.2.0
pillow==10.3.0
pillow-heif==0.16.0
watchdog==4.0.0
numpy==1.26.4
python-multipart==0.0.9
aiofiles==23.2.1
exifread==3.0.0
tqdm==4.66.2
hnswlib==0.8.0
open-clip-torch==2.24.0
torch>=2.0.0
